requirements.txt
streamlit>=1.30.0
anthropic>=0.40.0
firecrawl>=1.0.0
mem0ai>=0.1.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
